tools/player: switch text_last_{frame_index,duration} type from int to int64_t
duration is also adjusted to int64_t because it is derived from
p->duration (which is int64_t) and then assigned to text_last_duration.

p->frame_index is already int64_t so it makes sense to align
text_last_frame_index with it.

+++ ngl-tools/player.c
@@ -131,7 +131,7 @@ static void update_text(struct player *p)
         return;
 
     const int frame_ts = p->frame_time;
-    const int duration = p->duration / 1000000;
+    const int64_t duration = p->duration / 1000000;
     if (p->frame_index == p->text_last_frame_index && duration == p->text_last_duration)
         return;
